+++ torchbearer/callbacks/__init__.py
@@ -75,6 +75,13 @@ Weight Decay
         :members:
         :undoc-members:
 
+Weight / Bias Initialisation
+------------------------------------
+
+..  automodule:: torchbearer.callbacks.init
+        :members:
+        :undoc-members:
+
 Decorators
 ------------------------------------
 
@@ -94,6 +101,7 @@ from .tensor_board import TensorBoard, TensorBoardImages, TensorBoardProjector, 
 from .terminate_on_nan import *
 from .torch_scheduler import *
 from .weight_decay import *
+from . import init
 from .aggregate_predictions import *
 from .decorators import *
 from .live_loss_plot import LiveLossPlot